feat: implement Statement#expand() for table-prefixed columns
- Upgrade rusqlite from 0.32.1 to 0.38.0 for column_metadata feature
- Add column_tables tracking to Statement for expand mode
- Use columns_with_metadata() to get table origin for each column
- In expand mode, group result columns by their origin table:
  { tableName: { col1: val1, col2: val2 }, ... }
- Fix API changes in rusqlite 0.38: serialize(), load_extension(),
  create_scalar_function(), create_aggregate_function(), FromSql

Tests: 109 passing (was 107)

+++ noria-sqlite/src/dataflow/adapter.rs
@@ -71,10 +71,10 @@ impl SqliteAdapter {
         // Use PRAGMA table_info to get column information
         let mut stmt = conn.prepare(&format!("PRAGMA table_info({})", table_name))?;
         let column_iter = stmt.query_map([], |row| {
-            let cid: usize = row.get(0)?;
+            let cid: i64 = row.get(0)?;
             let name: String = row.get(1)?;
             let pk: bool = row.get(5)?;
-            Ok((cid, name, pk))
+            Ok((cid as usize, name, pk))
         })?;
 
         for col_result in column_iter {
